refactor(client): component folder structure
