+++ client/package.json
@@ -11,11 +11,14 @@
   },
   "dependencies": {
     "@emotion/react": "^11.9.0",
+    "@mantine/core": "^4.2.8",
+    "@mantine/hooks": "^4.2.8",
     "axios": "^0.27.2",
     "dayjs": "^1.11.2",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-query": "^3.39.1"
+    "react-query": "^3.39.1",
+    "tabler-icons-react": "^1.48.1"
   },
   "devDependencies": {
     "@emotion/babel-plugin": "^11.9.2",
