+++ client/src/App.tsx
@@ -1,6 +1,6 @@
-import { AppLayout } from '@components/app-layout'
-import { CreateTodoForm } from '@components/create-todo-form'
-import { TodoList } from '@components/todo-list'
+import { CreateTodoForm } from '@components/domain-component/create-todo-form'
+import { TodoList } from '@components/domain-component/todo-list'
+import { AppLayout } from '@components/layout-component/app-layout'
 import { MAX_CONTENT_WIDTH_PX_UNIT, MIN_CONTENT_WIDTH_PX_UNIT } from '@constants/style.constant'
 import { css } from '@emotion/react'
 import React from 'react'
